fixed schedule constructor, finished methods for hv, checked double return types in facade

+++ camp/Session.java
@@ -83,7 +83,7 @@ public class Session {
      * @param startDate the date the session starts
      * @param endDate the date the session ends
      */
-    public Session(String theme, int sessionNumber, Date startDate, Date endDate){
+    public Session(String theme, double sessionNumber, Date startDate, Date endDate){
         this.theme = theme;
         this.sessionNumber = sessionNumber;
         this.startDate = startDate;
